fix: contract benchmark C files use int64_t + @pure annotations
- All 7 C benchmark files: long/long long → int64_t (Windows 32-bit long overflow fix)
- purity_opt/bmb: @pure annotations → memory(none) → 2.88x FASTER vs Clang
- invariant_hoist/bmb: @pure on expensive_compute, scaled iterations 100→10000

+++ benches/contract/aliasing/c/main.c
@@ -3,47 +3,48 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 #define SIZE 1000
 
 // Without restrict: compiler must assume potential aliasing
-void process_arrays(long* a, long* b, int size) {
+void process_arrays(int64_t* a, int64_t* b, int size) {
     for (int i = 0; i < size; i++) {
-        long a_val = a[i];
+        int64_t a_val = a[i];
         b[i] = a_val + 1;       // Store to b
-        long a_val2 = a[i];     // Must reload - might alias with b
+        int64_t a_val2 = a[i];     // Must reload - might alias with b
         b[i] += a_val2;
     }
 }
 
 // With restrict: compiler can optimize (but not verified)
-void process_arrays_restrict(long* restrict a, long* restrict b, int size) {
+void process_arrays_restrict(int64_t* restrict a, int64_t* restrict b, int size) {
     for (int i = 0; i < size; i++) {
-        long a_val = a[i];
+        int64_t a_val = a[i];
         b[i] = a_val + 1;
-        long a_val2 = a[i];     // Can reuse cached value
+        int64_t a_val2 = a[i];     // Can reuse cached value
         b[i] += a_val2;
     }
 }
 
 // Vector addition with potential aliasing issues
-void vector_add(long* a, long* b, long* c, int size) {
+void vector_add(int64_t* a, int64_t* b, int64_t* c, int size) {
     for (int i = 0; i < size; i++) {
         c[i] = a[i] + b[i];  // If c aliases a or b, different result
     }
 }
 
 int main() {
-    long* a = (long*)malloc(SIZE * sizeof(long));
-    long* b = (long*)malloc(SIZE * sizeof(long));
-    long* c = (long*)malloc(SIZE * sizeof(long));
+    int64_t* a = (int64_t*)malloc(SIZE * sizeof(int64_t));
+    int64_t* b = (int64_t*)malloc(SIZE * sizeof(int64_t));
+    int64_t* c = (int64_t*)malloc(SIZE * sizeof(int64_t));
 
     for (int i = 0; i < SIZE; i++) {
         a[i] = i;
         b[i] = 0;
     }
 
-    long total = 0;
+    int64_t total = 0;
     for (int iter = 0; iter < 100; iter++) {
         process_arrays(a, b, SIZE);
         for (int i = 0; i < SIZE; i++) {
@@ -51,7 +52,7 @@ int main() {
         }
     }
 
-    printf("%ld\n", total);
+    printf("%lld\n", (long long)total);
 
     free(a);
     free(b);
